house/requirement.txt
Flask
Flask-SQLALchemy
PyMysql
redis
Flask-Script
Flask-Login
Flask-session
Flask-blueprint